requirements.txt
blinker==1.3
click==6.6
dominate==2.2.1
dropbox==6.6.0
facebook-sdk==1.0.0
Flask==0.11.1
Flask-Bootstrap==3.3.6.0
Flask-Login==0.2.9
flask-nav==0.5
Flask-Stormpath==0.4.5
Flask-WTF==0.12
gunicorn==19.6.0
httplib2==0.9.2
isodate==0.5.4
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauth2client==1.2
oauthlib==1.0.3
paho-mqtt==1.2
PyDispatcher==2.0.5
PyJWT==1.4.1
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
stormpath==2.1.1
urllib3==1.16
visitor==0.1.3
Werkzeug==0.11.10
WTForms==2.1
